Get win probability chart data from API

Get win probability chart data from API

cfb_metrics_wp(game_id, adjust_for_spread = TRUE)

Arguments

game_id
(Integer required): Game ID filter for querying a single game Can be found using the `cfb_game_info()` function

adjust_for_spread
(Logical default TRUE): Toggles pre-game spread adjustments

Examples


cfb_metrics_wp(game_id = 401012356)

#>                 playId
#> 1                    0
#> 2   401012356101855402
#> 3   401012356101855901
#> 4   401012356101856401
#> 5   401012356101859801
#> 6   401012356101866401
#> 7   401012356101867401
#> 8   401012356101875401
#> 9   401012356101879401
#> 10  401012356101884101
#> 11  401012356101885801
#> 12  401012356101889201
#> 13  401012356101894001
#> 14  401012356101895401
#> 15  401012356101899301
#> 16  401012356101906401
#> 17  401012356101914301
#> 18  401012356101916401
#> 19  401012356101919601
#> 20  401012356101925401
#> 21  401012356101926902
#> 22  401012356101927902
#> 23  401012356101936801
#> 24  401012356101938401
#> 25  401012356101945201
#> 26  401012356101947701
#> 27  401012356101955101
#> 28  401012356101958101
#> 29  401012356101959901
#> 30  401012356101966501
#> 31  401012356101969601
#> 32  401012356101974901
#> 33  401012356101976101
#> 34  401012356101979901
#> 35  401012356101988401
#> 36  401012356101995101
#> 37  401012356101995801
#> 38  401012356101996501
#> 39  401012356101999901
#> 40  401012356102854201
#> 41  401012356102856501
#> 42  401012356102858601
#> 43  401012356102866001
#> 44  401012356102867301
#> 45  401012356102876401
#> 46  401012356102879801
#> 47  401012356102888801
#> 48  401012356102896801
#> 49  401012356102897601
#> 50  401012356102904101
#> 51  401012356102908101
#> 52  401012356102915401
#> 53  401012356102915402
#> 54  401012356102916401
#> 55  401012356102919201
#> 56  401012356102927101
#> 57  401012356102935401
#> 58  401012356102938701
#> 59  401012356102946401
#> 60  401012356102946901
#> 61  401012356102948201
#> 62  401012356102956601
#> 63  401012356102964301
#> 64  401012356102966901
#> 65  401012356102975401
#> 66  401012356102978601
#> 67  401012356102984301
#> 68  401012356102984701
#> 69  401012356102985201
#> 70  401012356102985901
#> 71  401012356102986401
#> 72  401012356102987001
#> 73  401012356102987701
#> 74  401012356102989001
#> 75  401012356102989701
#> 76  401012356102994401
#> 77  401012356102996001
#> 78  401012356102996601
#> 79  401012356102997801
#> 80  401012356102999901
#> 81  401012356103855601
#> 82  401012356103858301
#> 83  401012356103859101
#> 84  401012356103864201
#> 85  401012356103865201
#> 86  401012356103868401
#> 87  401012356103876801
#> 88  401012356103878301
#> 89  401012356103884001
#> 90  401012356103885201
#> 91  401012356103888001
#> 92  401012356103895101
#> 93  401012356103899001
#> 94  401012356103905401
#> 95  401012356103906401
#> 96  401012356103908501
#> 97  401012356103914801
#> 98  401012356103916401
#> 99  401012356103925701
#> 100 401012356103926301
#> 101 401012356103926901
#> 102 401012356103927401
#> 103 401012356103928301
#> 104 401012356103929701
#> 105 401012356103936601
#> 106 401012356103937901
#> 107 401012356103938801
#> 108 401012356103947401
#> 109 401012356103955301
#> 110 401012356103959201
#> 111 401012356103966601
#> 112 401012356103969801
#> 113 401012356103976101
#> 114 401012356103985601
#> 115 401012356103987301
#> 116 401012356103988501
#> 117 401012356103996101
#> 118 401012356103999901
#> 119 401012356104855101
#> 120 401012356104856401
#> 121 401012356104857401
#> 122 401012356104864401
#> 123 401012356104866901
#> 124 401012356104875901
#> 125 401012356104877501
#> 126 401012356104879101
#> 127 401012356104886201
#> 128 401012356104894901
#> 129 401012356104898701
#> 130 401012356104899801
#> 131 401012356104907501
#> 132 401012356104909101
#> 133 401012356104915801
#> 134 401012356104917401
#> 135 401012356104924401
#> 136 401012356104929001
#> 137 401012356104935801
#> 138 401012356104937101
#> 139 401012356104944601
#> 140 401012356104948101
#> 141 401012356104955401
#> 142 401012356104957901
#> 143 401012356104964101
#> 144 401012356104965501
#> 145 401012356104968801
#> 146 401012356104974801
#> 147 401012356104975901
#> 148 401012356104976701
#> 149 401012356104977201
#> 150 401012356104977701
#> 151 401012356104978901
#> 152 401012356104987001
#> 153 401012356104987601
#> 154 401012356104988301
#> 155 401012356104989201
#> 156 401012356104994701
#> 157 401012356104996101
#> 158 401012356104997301
#> 159 401012356104997901
#> 160 401012356104998401
#> 161 401012356104998901
#> 162 401012356104999601
#> 163 401012356104999901
#>                                                                                                                                                                                  playText
#> 1                                                                                                                                                                              Game start
#> 2                                                                                                                    Joe Burrow pass complete to Lanard Fournette for 9 yds to the LSU 34
#> 3                                                                                                                                  Nick Brossette run for a loss of 2 yards to the LSU 32
#> 4                                                                                                                                                              Joe Burrow pass incomplete
#> 5                                                                                                         Zach Von Rosenberg punt for 47 yds, fair catch by Roshauud Paul at the TexAM 21
#> 6                                                                                                                                         Trayveon Williams run for 4 yds to the TexAM 25
#> 7                                                                                                     Kellen Mond pass complete to Jace Sternberger for a loss of 3 yards to the TexAM 22
#> 8                                                                                                                Kellen Mond sacked by Devin White for a loss of 12 yards to the TexAM 10
#> 9                                                                                                                                       Braden Mann punt for 53 yds, downed at the LSU 37
#> 10                                                                                                                   Joe Burrow pass complete to Stephen Sullivan for 9 yds to the LSU 46
#> 11                                                                                                                                 Nick Brossette run for a loss of 2 yards to the LSU 44
#> 12                                                                                                                                     Joe Burrow pass incomplete to Terrace Marshall Jr.
#> 13                                                                                                         Zach Von Rosenberg punt for 51 yds, fair catch by Roshauud Paul at the TexAM 5
#> 14                                                                                                                                         Trayveon Williams run for 3 yds to the TexAM 8
#> 15                                                                                                                   Kellen Mond pass complete to Jhamon Ausbon for 6 yds to the TexAM 14
#> 16                                                                                                                         Trayveon Williams run for 7 yds to the TexAM 21 for a 1ST down
#> 17                                                                                                                  Kellen Mond pass complete to Quartney Davis for 7 yds to the TexAM 28
#> 18                                                                                                  Kellen Mond pass complete to Quartney Davis for 13 yds to the TexAM 41 for a 1ST down
#> 19                                                                                                                                        Kellen Mond pass incomplete to Jace Sternberger
#> 20                                            Trayveon Williams run for 24 yds to the TEXASAM 35 for a 1ST down LSU Penalty, Face Mask (15 yards) (15 Yards) to the LSU 20 for a 1ST down
#> 21                                                                                                                                          Trayveon Williams run for 9 yds to the LSU 11
#> 22                                                                                                       Kellen Mond pass complete to Jhamon Ausbon for 1 yd to the LSU 10 for a 1ST down
#> 23                                                                                                                           Trayveon Williams run for 10 yds for a TD, (Seth Small KICK)
#> 24                                                                                                                       Joe Burrow pass complete to Dee Anderson for 8 yds to the LSU 33
#> 25                                                                                                                      Joe Burrow pass complete to Jonathan Giles for 1 yd to the LSU 34
#> 26                                                                                                                       Clyde Edwards-Helaire run for 3 yds to the LSU 37 for a 1ST down
#> 27                                                                                                                                      Clyde Edwards-Helaire run for 3 yds to the LSU 40
#> 28                                                                                                                           Joe Burrow run for 10 yds to the 50 yard line for a 1ST down
#> 29                                                                                                                         TEXAS A&M Penalty, Defensive Offside (5 Yards) to the TexAM 45
#> 30                                                                                                                        Clyde Edwards-Helaire run for a loss of 3 yards to the TexAM 48
#> 31                                                                                                   Joe Burrow pass complete to Jonathan Giles for 11 yds to the TexAM 37 for a 1ST down
#> 32                                                                                                                                         Joe Burrow pass incomplete to Justin Jefferson
#> 33                                                                                                                   Joe Burrow pass complete to Nick Brossette for 5 yds to the TexAM 32
#> 34                                                                                                                                Joe Burrow run for 7 yds to the TexAM 25 for a 1ST down
#> 35                                                                                                                                           Nick Brossette run for 3 yds to the TexAM 22
#> 36                                                                                                                                   Joe Burrow run for 22 yds for a TD (Cole Tracy KICK)
#> 37                                                                                                                                        Kellen Mond pass incomplete to Jace Sternberger
#> 38                                                                                                                        Trayveon Williams run for 20 yds to the TexAM 45 for a 1ST down
#> 39                                                                                                                                        Trayveon Williams run for 2 yds to the TexAM 47
#> 40                                                                                                                                        Kellen Mond pass incomplete to Jace Sternberger
#> 41                                                                                                     Kellen Mond pass complete to Jhamon Ausbon for 13 yds to the LSU 40 for a 1ST down
#> 42                                                                                                                                Kellen Mond run for 20 yds to the LSU 20 for a 1ST down
#> 43                                                                                                                                          Kellen Mond pass incomplete to Quartney Davis
#> 44                                                                                                                                                Kellen Mond run for 9 yds to the LSU 11
#> 45                                                                                                                                  Kellen Mond run for 1 yd to the LSU 10 for a 1ST down
#> 46                                                                                                                              Trayveon Williams run for a loss of 2 yards to the LSU 12
#> 47                                                                                                                                                Kellen Mond run for 2 yds to the LSU 10
#> 48                                                                                                    Kellen Mond pass complete to Jace Sternberger for 10 yds for a TD (Seth Small KICK)
#> 49                                                                                                      Joe Burrow pass complete to Foster Moreau for 13 yds to the LSU 38 for a 1ST down
#> 50                                                                                                    Joe Burrow pass complete to Ja'Marr Chase for 14 yds to the TexAM 48 for a 1ST down
#> 51                                                                                                                                               Joe Burrow run for 5 yds to the TexAM 43
#> 52                                                                                                                                            Joe Burrow pass incomplete to Ja'Marr Chase
#> 53                                                                                                                                    LSU Penalty, False Start (-5 Yards) to the TexAM 48
#> 54                                                                                                   Joe Burrow pass complete to Derrick Dillon for 10 yds to the TexAM 38 for a 1ST down
#> 55                                                                                                 Joe Burrow pass complete to Lanard Fournette for 19 yds to the TexAM 19 for a 1ST down
#> 56                                                                                                                                  LSU Penalty, Delay of Game (-5 Yards) to the TexAM 24
#> 57                                                                                                                 Joe Burrow sacked by Daylon Mack for a loss of 9 yards to the TexAM 33
#> 58                                                                                                                                               Joe Burrow run for 2 yds to the TexAM 31
#> 59                                                                                                                                             Joe Burrow pass incomplete to Dee Anderson
#> 60                                                                                                                                                             Cole Tracy 49 yd FG MISSED
#> 61                                                                                                                        Trayveon Williams run for 12 yds to the TexAM 43 for a 1ST down
#> 62                                                                                                                                        Trayveon Williams run for 4 yds to the TexAM 47
#> 63                                                                                                                                  Kellen Mond run for a loss of 2 yards to the TexAM 45
#> 64                                                                                                  Kellen Mond pass complete to Jace Sternberger for 36 yds to the LSU 19 for a 1ST down
#> 65                                                                                                                                          Trayveon Williams run for 3 yds to the LSU 16
#> 66                                                                                                                                             Kellen Mond pass incomplete to Trevor Wood
#> 67                                                                                                                                              Kellen Mond run for no gain to the LSU 16
#> 68                                                                                                                                                               Seth Small 33 yd FG GOOD
#> 69                                                                                                                                         Joe Burrow pass incomplete to Stephen Sullivan
#> 70                                                                                                                                         Joe Burrow pass incomplete to Justin Jefferson
#> 71                                                                                                      Joe Burrow pass complete to Ja'Marr Chase for 11 yds to the LSU 46 for a 1ST down
#> 72                                                                                                                                               Joe Burrow run for 7 yds to the TexAM 47
#> 73                                                                                                     Joe Burrow pass complete to Ja'Marr Chase for 6 yds to the TexAM 41 for a 1ST down
#> 74                                                                                                                                         Joe Burrow pass incomplete to Justin Jefferson
#> 75                                                                                                                 Joe Burrow pass complete to Justin Jefferson for 4 yds to the TexAM 37
#> 76                                                                                                                                Joe Burrow run for 7 yds to the TexAM 30 for a 1ST down
#> 77                                                                                                                                         Joe Burrow pass incomplete to Stephen Sullivan
#> 78                                                                                                                                             Joe Burrow pass incomplete to Dee Anderson
#> 79                                                                                                                                                Joe Burrow run for 1 yd to the TexAM 29
#> 80                                                                                                                                                               Cole Tracy 47 yd FG GOOD
#> 81                                                                                                                                        Trayveon Williams run for 2 yds to the TexAM 27
#> 82                                                                                                                                           Kellen Mond pass incomplete to Jhamon Ausbon
#> 83                                                                                                                                           Kellen Mond pass incomplete to Jhamon Ausbon
#> 84                                                                                               Braden Mann punt for 59 yds , Jonathan Giles returns for a loss of 4 yards to the LSU 10
#> 85                                                                                                                                             Nick Brossette run for 9 yds to the LSU 19
#> 86                                                                                                                              Nick Brossette run for 7 yds to the LSU 26 for a 1ST down
#> 87                                                                                             Joe Burrow pass complete to Justin Jefferson for 24 yds to the 50 yard line for a 1ST down
#> 88                                                                                                                                         Joe Burrow pass incomplete to Lanard Fournette
#> 89                                                                                                                   Joe Burrow pass complete to Derrick Dillon for 6 yds to the TexAM 44
#> 90                                                                                              Joe Burrow pass complete to Terrace Marshall Jr. for 8 yds to the TexAM 36 for a 1ST down
#> 91                                                                                                                   Joe Burrow pass complete to Nick Brossette for 2 yds to the TexAM 34
#> 92                                                                                                                                             Joe Burrow run for no gain to the TexAM 34
#> 93                                                                                                                               Joe Burrow run for 12 yds to the TexAM 22 for a 1ST down
#> 94                                                                                                                                           Nick Brossette run for 9 yds to the TexAM 13
#> 95                                                                                                                                         Nick Brossette run for no gain to the TexAM 13
#> 96                                                                                                                                Joe Burrow run for 2 yds to the TexAM 11 for a 1ST down
#> 97                                                                                                                                Nick Brossette run for a loss of 1 yard to the TexAM 12
#> 98                                                                                                                                                Joe Burrow run for 5 yds to the TexAM 7
#> 99                                                                                                      Joe Burrow pass complete to Justin Jefferson for 7 yds for a TD (Cole Tracy KICK)
#> 100                                                                                                                                        Kellen Mond pass incomplete to Kendrick Rogers
#> 101                                                                                                                                         Kellen Mond pass incomplete to Quartney Davis
#> 102                                                                                                                                         Kellen Mond pass incomplete to Quartney Davis
#> 103                                                             Braden Mann punt for 46 yds , Jonathan Giles returns for no gain to the LSU 29 Jonathan Giles fumbled, recovered by TexAM
#> 104                                                                                                                                         Trayveon Williams run for 4 yds to the LSU 25
#> 105                                                                                                                          Trayveon Williams run for 6 yds to the LSU 19 for a 1ST down
#> 106                                                                                                                                          Kellen Mond pass incomplete to Jhamon Ausbon
#> 107                                                                                                                                         Trayveon Williams run for 3 yds to the LSU 16
#> 108                                                                                                                                 Kellen Mond run for 8 yds to the LSU 8 for a 1ST down
#> 109                                                                                                                                          Trayveon Williams run for 3 yds to the LSU 5
#> 110                                                                                                                                          Trayveon Williams run for 4 yds to the LSU 1
#> 111                                                                                                                             Trayveon Williams run for 1 yd for a TD (Seth Small KICK)
#> 112                                                                                                               Joe Burrow sacked by Debione Renfro for a loss of 8 yards to the LSU 18
#> 113                                                                                                                     Joe Burrow pass complete to Foster Moreau for 7 yds to the LSU 25
#> 114                                                                                                                Joe Burrow sacked by Tyree Johnson for a loss of 3 yards to the LSU 22
#> 115                                                                                       Zach Von Rosenberg punt for 53 yds , Roshauud Paul returns for a loss of 1 yard to the TexAM 24
#> 116                                                                                                                                        Trayveon Williams run for 1 yd to the TexAM 25
#> 117                                                                                                                                         Kellen Mond pass incomplete to Camron Buckley
#> 118                                                                                                          Kellen Mond sacked by Rashard Lawrence for a loss of 7 yards to the TexAM 18
#> 119                                                      Braden Mann punt for 56 yds , Jonathan Giles returns for 3 yds to the LSU 21 LSU Penalty, Illegal Block (10 Yards) to the LSU 19
#> 120                                                                                                                            Nick Brossette run for 16 yds to the LSU 35 for a 1ST down
#> 121                                                                                                  Joe Burrow pass complete to Justin Jefferson for 11 yds to the LSU 46 for a 1ST down
#> 122                                                                                                                                      Nick Brossette run for 4 yds to the 50 yard line
#> 123                                                                                                                  Joe Burrow sacked by Daylon Mack for a loss of 3 yards to the LSU 47
#> 124                                                                                                                  Joe Burrow sacked by Otaro Alaka for a loss of 3 yards to the LSU 44
#> 125                                    Zach Von Rosenberg punt for 37 yds, fair catch by Roshauud Paul at the TEXASAM 19 LSU Penalty, Unsportsmanlike Conduct (-15 Yards) to the TexAM 34
#> 126                                                                                                                                       Trayveon Williams run for 4 yds to the TexAM 38
#> 127                                                                                                                                       Trayveon Williams run for 4 yds to the TexAM 42
#> 128                                                                                                 Kellen Mond pass complete to Cullen Gillaspia for 18 yds to the LSU 40 for a 1ST down
#> 129 Kellen Mond pass complete to Trayveon Williams for a loss of 2 yards for a TD Trayveon Williams fumbled, recovered by LSU Michael Divinity Jr.  , return for 58 yds (Cole Tracy KICK)
#> 130                                                                                                                              Jashaun Corbin run for a loss of 6 yards to the TexAM 19
#> 131                                                                                                                                      Kellen Mond pass incomplete to Trayveon Williams
#> 132                                                                                                                                       Trayveon Williams run for 5 yds to the TexAM 24
#> 133                                                                                                               Braden Mann punt for 52 yds, fair catch by Jonathan Giles at the LSU 24
#> 134                                                                                                                        Lanard Fournette run for 46 yds to the TexAM 30 for a 1ST down
#> 135                                                                                                                                          Nick Brossette run for 9 yds to the TexAM 21
#> 136                                                                                                                           Nick Brossette run for 7 yds to the TexAM 14 for a 1ST down
#> 137                                                                                                       Joe Burrow pass complete to Foster Moreau for 14 yds for a TD (Cole Tracy KICK)
#> 138                                                                                                               Kellen Mond pass complete to Jace Sternberger for 7 yds to the TexAM 32
#> 139                                                                                                                        Trayveon Williams run for 3 yds to the TexAM 35 for a 1ST down
#> 140                                                                                                                                             Kellen Mond run for 4 yds to the TexAM 39
#> 141                                                                                                   Kellen Mond pass complete to Jhamon Ausbon for 8 yds to the TexAM 47 for a 1ST down
#> 142                                                                                                                                       Kellen Mond pass incomplete to Jace Sternberger
#> 143                                                                                                                    Kellen Mond pass complete to Jhamon Ausbon for 9 yds to the LSU 44
#> 144                                                                                                                          Trayveon Williams run for 2 yds to the LSU 42 for a 1ST down
#> 145                                                                                                                   Kellen Mond pass complete to Camron Buckley for 5 yds to the LSU 37
#> 146                                                                                                                                          Kellen Mond pass incomplete to Jhamon Ausbon
#> 147                                                                                                                                             Kellen Mond run for no gain to the LSU 37
#> 148                                                                                                                                               Kellen Mond run for 3 yds to the LSU 34
#> 149                                                                                                                                             Nick Brossette run for 1 yd to the LSU 35
#> 150                                                                                                                                                Joe Burrow run for 5 yds to the LSU 40
#> 151                                                                                                                                     Joe Burrow run for a loss of 1 yard to the LSU 39
#> 152                                                                                                       Zach Von Rosenberg punt for 39 yds, fair catch by Roshauud Paul at the TexAM 22
#> 153                                                                                                                                         Kellen Mond pass incomplete to Camron Buckley
#> 154                                                                                                                                         Kellen Mond pass incomplete to Quartney Davis
#> 155                                                                                                 Kellen Mond pass complete to Quartney Davis for 12 yds to the TexAM 34 for a 1ST down
#> 156                                                                                                 Kellen Mond pass complete to Quartney Davis for 13 yds to the TexAM 47 for a 1ST down
#> 157                                                                                                                                        Kellen Mond pass incomplete to Kendrick Rogers
#> 158                                                                                         Kellen Mond run for a loss of 8 yards to the TexAM 39 Kellen Mond fumbled, recovered by TexAM
#> 159                                                                                                                                       Kellen Mond pass incomplete to Jace Sternberger
#> 160                                                                                                   Kellen Mond pass complete to Quartney Davis for 20 yds to the LSU 41 for a 1ST down
#> 161                                                                                                                                                                  TEAM pass incomplete
#> 162                                                                                                  Kellen Mond pass complete to Kendrick Rogers for 22 yds to the LSU 19 for a 1ST down
#> 163                                                                                                     Kellen Mond pass complete to Quartney Davis for 19 yds for a TD (Seth Small KICK)
#>     homeId      home awayId away spread homeBall homeScore awayScore
#> 1      245 Texas A&M     99  LSU      0    FALSE         0         0
#> 2      245 Texas A&M     99  LSU      0    FALSE         0         0
#> 3      245 Texas A&M     99  LSU      0    FALSE         0         0
#> 4      245 Texas A&M     99  LSU      0    FALSE         0         0
#> 5      245 Texas A&M     99  LSU      0    FALSE         0         0
#> 6      245 Texas A&M     99  LSU      0     TRUE         0         0
#> 7      245 Texas A&M     99  LSU      0     TRUE         0         0
#> 8      245 Texas A&M     99  LSU      0     TRUE         0         0
#> 9      245 Texas A&M     99  LSU      0     TRUE         0         0
#> 10     245 Texas A&M     99  LSU      0    FALSE         0         0
#> 11     245 Texas A&M     99  LSU      0    FALSE         0         0
#> 12     245 Texas A&M     99  LSU      0    FALSE         0         0
#> 13     245 Texas A&M     99  LSU      0    FALSE         0         0
#> 14     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 15     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 16     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 17     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 18     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 19     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 20     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 21     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 22     245 Texas A&M     99  LSU      0     TRUE         0         0
#> 23     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 24     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 25     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 26     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 27     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 28     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 29     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 30     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 31     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 32     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 33     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 34     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 35     245 Texas A&M     99  LSU      0    FALSE         7         0
#> 36     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 37     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 38     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 39     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 40     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 41     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 42     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 43     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 44     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 45     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 46     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 47     245 Texas A&M     99  LSU      0     TRUE         7         7
#> 48     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 49     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 50     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 51     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 52     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 53     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 54     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 55     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 56     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 57     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 58     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 59     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 60     245 Texas A&M     99  LSU      0    FALSE        14         7
#> 61     245 Texas A&M     99  LSU      0     TRUE        14         7
#> 62     245 Texas A&M     99  LSU      0     TRUE        14         7
#> 63     245 Texas A&M     99  LSU      0     TRUE        14         7
#> 64     245 Texas A&M     99  LSU      0     TRUE        14         7
#> 65     245 Texas A&M     99  LSU      0     TRUE        14         7
#> 66     245 Texas A&M     99  LSU      0     TRUE        14         7
#> 67     245 Texas A&M     99  LSU      0     TRUE        14         7
#> 68     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 69     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 70     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 71     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 72     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 73     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 74     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 75     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 76     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 77     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 78     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 79     245 Texas A&M     99  LSU      0    FALSE        17         7
#> 80     245 Texas A&M     99  LSU      0     TRUE        17        10
#> 81     245 Texas A&M     99  LSU      0     TRUE        17        10
#> 82     245 Texas A&M     99  LSU      0     TRUE        17        10
#> 83     245 Texas A&M     99  LSU      0     TRUE        17        10
#> 84     245 Texas A&M     99  LSU      0     TRUE        17        10
#> 85     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 86     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 87     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 88     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 89     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 90     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 91     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 92     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 93     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 94     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 95     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 96     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 97     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 98     245 Texas A&M     99  LSU      0    FALSE        17        10
#> 99     245 Texas A&M     99  LSU      0     TRUE        17        17
#> 100    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 101    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 102    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 103    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 104    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 105    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 106    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 107    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 108    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 109    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 110    245 Texas A&M     99  LSU      0     TRUE        17        17
#> 111    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 112    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 113    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 114    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 115    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 116    245 Texas A&M     99  LSU      0     TRUE        24        17
#> 117    245 Texas A&M     99  LSU      0     TRUE        24        17
#> 118    245 Texas A&M     99  LSU      0     TRUE        24        17
#> 119    245 Texas A&M     99  LSU      0     TRUE        24        17
#> 120    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 121    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 122    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 123    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 124    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 125    245 Texas A&M     99  LSU      0    FALSE        24        17
#> 126    245 Texas A&M     99  LSU      0     TRUE        24        17
#> 127    245 Texas A&M     99  LSU      0     TRUE        24        17
#> 128    245 Texas A&M     99  LSU      0     TRUE        24        17
#> 129    245 Texas A&M     99  LSU      0    FALSE        24        24
#> 130    245 Texas A&M     99  LSU      0     TRUE        24        24
#> 131    245 Texas A&M     99  LSU      0     TRUE        24        24
#> 132    245 Texas A&M     99  LSU      0     TRUE        24        24
#> 133    245 Texas A&M     99  LSU      0     TRUE        24        24
#> 134    245 Texas A&M     99  LSU      0    FALSE        24        24
#> 135    245 Texas A&M     99  LSU      0    FALSE        24        24
#> 136    245 Texas A&M     99  LSU      0    FALSE        24        24
#> 137    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 138    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 139    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 140    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 141    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 142    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 143    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 144    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 145    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 146    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 147    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 148    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 149    245 Texas A&M     99  LSU      0    FALSE        24        31
#> 150    245 Texas A&M     99  LSU      0    FALSE        24        31
#> 151    245 Texas A&M     99  LSU      0    FALSE        24        31
#> 152    245 Texas A&M     99  LSU      0    FALSE        24        31
#> 153    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 154    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 155    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 156    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 157    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 158    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 159    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 160    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 161    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 162    245 Texas A&M     99  LSU      0     TRUE        24        31
#> 163    245 Texas A&M     99  LSU      0    FALSE        31        31
#>     timeRemaining yardLine down distance homeWinProb playNumber
#> 1            3600       65    1       10  0.48320506          0
#> 2            3585       75    1       10  0.46924645          1
#> 3            3580       66    2        1  0.48748890          2
#> 4            3575       68    3        3  0.44026201          3
#> 5            3541       68    4        3  0.39502371          4
#> 6            3515       79    1       10  0.51564712          5
#> 7            3505       75    2        6  0.49876953          6
#> 8            3465       78    3        9  0.42936535          7
#> 9            3425       90    4       21  0.28770530          8
#> 10           3418       63    1       10  0.48657118          9
#> 11           3401       54    2        1  0.50461351         10
#> 12           3367       56    3        3  0.47027795         11
#> 13           3359       56    4        3  0.42959267         12
#> 14           3345       95    1       10  0.47935462         13
#> 15           3306       92    2        7  0.45407570         14
#> 16           3275       86    3        1  0.45048126         15
#> 17           3236       79    1       10  0.51905846         16
#> 18           3215       72    2        3  0.52211830         17
#> 19           3183       59    1       10  0.56062782         18
#> 20           3165       35    2       10  0.57071064         19
#> 21           3150       20    1       10  0.61994760         20
#> 22           3140       11    2        1  0.62831300         21
#> 23           3091       10    1       10  0.59731020         22
#> 24           3075       75    1       10  0.67596381         23
#> 25           3047       67    2        2  0.68081326         24
#> 26           3022       66    3        1  0.67412871         25
#> 27           2988       63    1       10  0.67577590         26
#> 28           2958       60    2        7  0.67757587         27
#> 29           2940       50    1       10  0.67075174         28
#> 30           2914       45    1        5  0.66188565         29
#> 31           2883       48    2        8  0.67798407         30
#> 32           2870       37    1       10  0.65645433         31
#> 33           2858       37    2       10  0.67614880         32
#> 34           2820       32    3        5  0.68092010         33
#> 35           2775       25    1       10  0.63002696         34
#> 36           2748       22    2        7  0.61023557         35
#> 37           2741       75    1       10  0.54806420         36
#> 38           2734       75    2       10  0.50984968         37
#> 39           2700       55    1       10  0.58284701         38
#> 40           2697       53    2        8  0.56166267         39
#> 41           2674       53    3        8  0.52624347         40
#> 42           2653       40    1       10  0.60426747         41
#> 43           2619       20    1       10  0.62705513         42
#> 44           2606       20    2       10  0.60639333         43
#> 45           2555       11    3        1  0.61499041         44
#> 46           2521       10    1       10  0.63621924         45
#> 47           2471       12    2       12  0.61214525         46
#> 48           2431       10    3       10  0.67441078         47
#> 49           2423       75    1       10  0.67121668         48
#> 50           2398       62    1       10  0.67024949         49
#> 51           2358       48    1       10  0.66268589         50
#> 52           2325       43    2        5  0.67300179         51
#> 53           2325       43    3        5  0.67620686         52
#> 54           2315       48    3       10  0.67116113         53
#> 55           2287       38    1       10  0.65027626         54
#> 56           2248       19    1       10  0.60601750         55
#> 57           2205       24    1       15  0.63350876         56
#> 58           2172       33    2       24  0.66798165         57
#> 59           2135       31    3       22  0.66761575         58
#> 60           2130       31    4       22  0.65955909         59
#> 61           2117       69    1       10  0.72118789         60
#> 62           2073       57    1       10  0.72846848         61
#> 63           2036       53    2        6  0.72324442         62
#> 64           2010       55    3        8  0.70710321         63
#> 65           1965       19    1       10  0.75024035         64
#> 66           1933       16    2        7  0.74360871         65
#> 67           1916       16    3        7  0.73213177         66
#> 68           1912       16    4        7  0.72890228         67
#> 69           1907       65    1       10  0.72268990         68
#> 70           1900       65    2       10  0.72280466         69
#> 71           1895       65    3       10  0.71746383         70
#> 72           1889       54    1       10  0.72009370         71
#> 73           1882       47    2        3  0.72665985         72
#> 74           1869       41    1       10  0.71162785         73
#> 75           1862       41    2       10  0.72518042         74
#> 76           1855       37    3        6  0.72830934         75
#> 77           1839       30    1       10  0.69737711         76
#> 78           1833       30    2       10  0.72249699         77
#> 79           1821       30    3       10  0.72792812         78
#> 80           1800       29    4        9  0.71115057         79
#> 81           1783       75    1       10  0.71706286         80
#> 82           1756       73    2        8  0.70952739         81
#> 83           1748       73    3        8  0.69794101         82
#> 84           1737       73    4        8  0.68525921         83
#> 85           1727       90    1       10  0.67167410         84
#> 86           1695       81    2        1  0.67637905         85
#> 87           1651       74    1       10  0.67384802         86
#> 88           1636       50    1       10  0.66191261         87
#> 89           1619       50    2       10  0.67660175         88
#> 90           1607       44    3        4  0.68080148         89
#> 91           1579       36    1       10  0.64226588         90
#> 92           1548       34    2        8  0.67008987         91
#> 93           1509       34    3        8  0.68240960         92
#> 94           1485       22    1       10  0.60519549         93
#> 95           1475       13    2        1  0.61933306         94
#> 96           1454       13    3        1  0.66885045         95
#> 97           1431       11    1       10  0.55778274         96
#> 98           1415       12    2       11  0.64819117         97
#> 99           1362        7    3        6  0.62379347         98
#> 100          1356       75    1       10  0.58138350         99
#> 101          1350       75    2       10  0.55202458        100
#> 102          1345       75    3       10  0.51649425        101
#> 103          1336       75    4       10  0.47498554        102
#> 104          1322       29    1       10  0.63094240        103
#> 105          1293       25    2        6  0.62481006        104
#> 106          1280       19    1       10  0.63815434        105
#> 107          1271       19    2       10  0.62394916        106
#> 108          1225       16    3        7  0.61428478        107
#> 109          1186        8    1        8  0.64651307        108
#> 110          1147        5    2        5  0.64073568        109
#> 111          1113        1    3        1  0.66992464        110
#> 112          1081       74    1       10  0.70336497        111
#> 113          1058       82    2       18  0.70566519        112
#> 114          1003       75    3       11  0.71251927        113
#> 115           986       78    4       14  0.70579884        114
#> 116           974       76    1       10  0.75931276        115
#> 117           938       75    2        9  0.76015379        116
#> 118           900       75    3        9  0.76145086        117
#> 119           888       82    4       16  0.75291476        118
#> 120           875       81    1       10  0.72862303        119
#> 121           865       65    1       10  0.72354166        120
#> 122           835       54    1       10  0.71889817        121
#> 123           810       50    2        6  0.73367703        122
#> 124           760       53    3        9  0.75089437        123
#> 125           744       56    4       12  0.75386544        124
#> 126           728       66    1       10  0.79907944        125
#> 127           697       62    2        6  0.80353119        126
#> 128           650       58    3        2  0.81160462        127
#> 129           612       40    1       10  0.49231197        128
#> 130           601       75    1       10  0.60183638        129
#> 131           564       81    2       16  0.55339003        130
#> 132           548       81    3       16  0.51709512        131
#> 133           521       76    4       11  0.50336466        132
#> 134           505       76    1       10  0.50241923        133
#> 135           475       30    1       10  0.44487478        134
#> 136           429       21    2        1  0.44015111        135
#> 137           401       14    1       10  0.26573437        136
#> 138           388       75    1       10  0.15598737        137
#> 139           353       68    2        3  0.13729260        138
#> 140           318       65    1       10  0.15384176        139
#> 141           285       61    2        6  0.12485591        140
#> 142           260       53    1       10  0.15626025        141
#> 143           238       53    2       10  0.11707313        142
#> 144           224       44    3        1  0.10997625        143
#> 145           191       42    1       10  0.15245292        144
#> 146           171       37    2        5  0.12924440        145
#> 147           160       37    3        5  0.09810945        146
#> 148           152       37    4        5  0.07344420        147
#> 149           147       66    1       10  0.03936474        148
#> 150           142       65    2        9  0.03327678        149
#> 151           130       60    3        4  0.02910300        150
#> 152            89       61    4        5  0.02061695        151
#> 153            83       78    1       10  0.08599370        152
#> 154            76       78    2       10  0.06420434        153
#> 155            67       78    3       10  0.04759514        154
#> 156            52       66    1       10  0.09187904        155
#> 157            38       53    1       10  0.10202150        156
#> 158            26       53    2       10  0.07655120        157
#> 159            20       61    3       18  0.04625473        158
#> 160            15       61    4       18  0.03481416        159
#> 161            10       41    1       10  0.10819105        160
#> 162             3       41    2       10  0.08302641        161
#> 163             0       19    2       10  1.00000000        162