Gets player info associated by play

Information describes the players involved in the play this includes passer, receiver, defensive players who create sacks or picks, etc

cfb_play_stats_player(
  year = NULL,
  week = NULL,
  team = NULL,
  game_id = NULL,
  athlete_id = NULL,
  stat_type_id = NULL,
  season_type = "regular"
)

Arguments

year
(Integer optional): Year, 4 digit format (YYYY)

week
(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

team
(String optional): D-I Team

game_id
(Integer optional): Game ID filter for querying a single game Can be found using the `cfb_game_info()` function

athlete_id
(Integer optional): Athlete ID filter for querying a single athlete Can be found using the `cfb_player_info()` function.

stat_type_id
(Integer optional): Stat Type ID filter for querying a single stat type Can be found using the `cfb_play_stats_types()` function

season_type
(String default regular): Select Season Type: regular, postseason, or both

Examples


ath_id <- as.numeric(cfb_player_info(search_term = 'terry', team = "florida state")$id)
cfb_play_stats_player(year = 2019, week = 1, team = 'Florida State', athlete_id = ath_id)

#>        gameId season week    opponent teamScore opponentScore     driveId
#> 1   401112431   2019    1 Boise State        14             6  4011124316
#> 2   401112431   2019    1 Boise State        14             6  4011124316
#> 3   401112431   2019    1 Boise State         0             0  4011124312
#> 4   401112431   2019    1 Boise State        14             6  4011124316
#> 5   401112431   2019    1 Boise State        14             6  4011124316
#> 6   401112431   2019    1 Boise State         0             0  4011124312
#> 7   401112431   2019    1 Boise State        14             6  4011124316
#> 8   401112431   2019    1 Boise State        14             6  4011124316
#> 9   401112431   2019    1 Boise State        14             3  4011124314
#> 10  401112431   2019    1 Boise State        14             3  4011124314
#> 11  401112431   2019    1 Boise State        14             3  4011124314
#> 12  401112431   2019    1 Boise State        14             3  4011124314
#> 13  401112431   2019    1 Boise State        14             6  4011124316
#> 14  401112431   2019    1 Boise State         0             0  4011124312
#> 15  401112431   2019    1 Boise State         0             0  4011124312
#> 16  401112431   2019    1 Boise State        14             6  4011124316
#> 17  401112431   2019    1 Boise State        14             6  4011124316
#> 18  401112431   2019    1 Boise State         7             0  4011124312
#> 19  401112431   2019    1 Boise State         7             0  4011124312
#> 20  401112431   2019    1 Boise State        14             6  4011124316
#> 21  401112431   2019    1 Boise State        14             6  4011124316
#> 22  401112431   2019    1 Boise State        14             6  4011124316
#> 23  401112431   2019    1 Boise State        14             6  4011124316
#> 24  401112431   2019    1 Boise State        14             6  4011124316
#> 25  401112431   2019    1 Boise State        21             6  4011124316
#> 26  401112431   2019    1 Boise State        21             6  4011124316
#> 27  401112431   2019    1 Boise State        21             6  4011124316
#> 28  401112431   2019    1 Boise State        21             6  4011124316
#> 29  401112431   2019    1 Boise State        21             6  4011124318
#> 30  401112431   2019    1 Boise State        21             6  4011124318
#> 31  401112431   2019    1 Boise State        31            13 40111243113
#> 32  401112431   2019    1 Boise State        21             6  4011124318
#> 33  401112431   2019    1 Boise State        21             6  4011124318
#> 34  401112431   2019    1 Boise State        21             6  4011124318
#> 35  401112431   2019    1 Boise State        24             6 40111243110
#> 36  401112431   2019    1 Boise State        21             6  4011124317
#> 37  401112431   2019    1 Boise State        24             6 40111243110
#> 38  401112431   2019    1 Boise State        21             6  4011124318
#> 39  401112431   2019    1 Boise State        24             6 40111243110
#> 40  401112431   2019    1 Boise State        24             6 40111243110
#> 41  401112431   2019    1 Boise State        21             6  4011124318
#> 42  401112431   2019    1 Boise State        24             6 40111243110
#> 43  401112431   2019    1 Boise State        31            16 40111243114
#> 44  401112431   2019    1 Boise State        31            16 40111243114
#> 45  401112431   2019    1 Boise State        21             6  4011124318
#> 46  401112431   2019    1 Boise State        24            13 40111243112
#> 47  401112431   2019    1 Boise State        24            13 40111243112
#> 48  401112431   2019    1 Boise State        24             6 40111243110
#> 49  401112431   2019    1 Boise State        24             6 40111243110
#> 50  401112431   2019    1 Boise State        31            16 40111243114
#> 51  401112431   2019    1 Boise State        31            16 40111243114
#> 52  401112431   2019    1 Boise State        31            16 40111243114
#> 53  401112431   2019    1 Boise State        31            13 40111243112
#> 54  401112431   2019    1 Boise State        31            13 40111243112
#> 55  401112431   2019    1 Boise State        31            13 40111243112
#> 56  401112431   2019    1 Boise State        31            13 40111243112
#> 57  401112431   2019    1 Boise State        24             6  4011124319
#> 58  401112431   2019    1 Boise State        21             6  4011124318
#> 59  401112431   2019    1 Boise State        21             6  4011124318
#> 60  401112431   2019    1 Boise State        31            19 40111243118
#> 61  401112431   2019    1 Boise State        31            19 40111243116
#> 62  401112431   2019    1 Boise State        31            19 40111243116
#> 63  401112431   2019    1 Boise State        31            19 40111243116
#> 64  401112431   2019    1 Boise State        31            19 40111243122
#> 65  401112431   2019    1 Boise State        31            19 40111243116
#> 66  401112431   2019    1 Boise State        31            19 40111243116
#> 67  401112431   2019    1 Boise State        31            26 40111243124
#> 68  401112431   2019    1 Boise State        31            26 40111243124
#> 69  401112431   2019    1 Boise State        31            19 40111243122
#> 70  401112431   2019    1 Boise State        31            19 40111243122
#> 71  401112431   2019    1 Boise State        31            19 40111243116
#> 72  401112431   2019    1 Boise State        31            19 40111243116
#> 73  401112431   2019    1 Boise State        31            26 40111243124
#> 74  401112431   2019    1 Boise State        31            26 40111243124
#> 75  401112431   2019    1 Boise State        31            19 40111243120
#> 76  401112431   2019    1 Boise State        31            19 40111243120
#> 77  401112431   2019    1 Boise State        31            19 40111243120
#> 78  401112431   2019    1 Boise State        31            19 40111243120
#> 79  401112431   2019    1 Boise State        31            19 40111243118
#> 80  401112431   2019    1 Boise State        31            19 40111243116
#> 81  401112431   2019    1 Boise State        31            19 40111243120
#> 82  401112431   2019    1 Boise State        31            19 40111243120
#> 83  401112431   2019    1 Boise State        31            19 40111243119
#> 84  401112431   2019    1 Boise State        31            19 40111243116
#> 85  401112431   2019    1 Boise State        31            19 40111243116
#> 86  401112431   2019    1 Boise State        31            26 40111243124
#> 87  401112431   2019    1 Boise State        31            19 40111243120
#> 88  401112431   2019    1 Boise State        31            19 40111243120
#> 89  401112431   2019    1 Boise State        31            19 40111243118
#> 90  401112431   2019    1 Boise State        31            19 40111243118
#> 91  401112431   2019    1 Boise State        31            19 40111243121
#> 92  401112431   2019    1 Boise State        31            19 40111243121
#> 93  401112431   2019    1 Boise State        31            33 40111243126
#> 94  401112431   2019    1 Boise State        31            33 40111243126
#> 95  401112431   2019    1 Boise State        31            36 40111243130
#> 96  401112431   2019    1 Boise State        31            36 40111243130
#> 97  401112431   2019    1 Boise State        31            36 40111243128
#> 98  401112431   2019    1 Boise State        31            36 40111243129
#> 99  401112431   2019    1 Boise State        31            33 40111243126
#> 100 401112431   2019    1 Boise State        31            33 40111243126
#> 101 401112431   2019    1 Boise State        31            36 40111243130
#> 102 401112431   2019    1 Boise State        31            36 40111243130
#> 103 401112431   2019    1 Boise State        31            33 40111243126
#> 104 401112431   2019    1 Boise State        31            36 40111243130
#> 105 401112431   2019    1 Boise State        31            36 40111243128
#> 106 401112431   2019    1 Boise State        31            36 40111243130
#> 107 401112431   2019    1 Boise State        31            36 40111243130
#> 108 401112431   2019    1 Boise State        31            36 40111243128
#> 109 401112431   2019    1 Boise State        31            36 40111243128
#>                 playId period secondsRemaining yardsToGoal down distance
#> 1   401112431101974101      1               58          77    2       11
#> 2   401112431101974101      1               58          77    2       11
#> 3   401112431101874401      1               55          47    1       10
#> 4   401112431101984601      1               53          38    1       10
#> 5   401112431101984601      1               53          38    1       10
#> 6   401112431101875401      1               45          47    2       10
#> 7   401112431101985601      1               43          37    2        9
#> 8   401112431101985601      1               43          37    2        9
#> 9   401112431101925701      1               42          75    1       10
#> 10  401112431101925701      1               42          75    1       10
#> 11  401112431101925701      1               42          75    1       10
#> 12  401112431101925701      1               42          75    1       10
#> 13  401112431101976101      1               38          65    1       10
#> 14  401112431101876801      1               31          47    3       10
#> 15  401112431101876801      1               31          47    3       10
#> 16  401112431101986801      1               31          25    1       10
#> 17  401112431101986802      1               31          15    1       10
#> 18  401112431101877101      1               28          38    4        1
#> 19  401112431101877101      1               28          38    4        1
#> 20  401112431101977401      1               25          51    1       10
#> 21  401112431101977401      1               25          51    1       10
#> 22  401112431101967401      1               25          81    1       15
#> 23  401112431101978701      1               12          51    2       10
#> 24  401112431101978701      1               12          51    2       10
#> 25  401112431101989401      1                5           5    1        5
#> 26  401112431101989401      1                5           5    1        5
#> 27  401112431101989401      1                5           5    1        5
#> 28  401112431101989401      1                5           5    1        5
#> 29  401112431102864001      2               59          89    1       10
#> 30  401112431102864001      2               59          89    1       10
#> 31  401112431102974201      2               57          20    3       10
#> 32  401112431102864501      2               54          75    1       10
#> 33  401112431102884601      2               53           2    2        2
#> 34  401112431102884601      2               53           2    2        2
#> 35  401112431102924801      2               51          32    2       10
#> 36  401112431102854901      2               50          70    2        8
#> 37  401112431102915501      2               44          68    1       10
#> 38  401112431102865701      2               42          75    2       10
#> 39  401112431102925701      2               42          29    3        7
#> 40  401112431102925701      2               42          29    3        7
#> 41  401112431102876201      2               37           1    1        1
#> 42  401112431102916201      2               37          63    2        5
#> 43  401112431102976701      2               32          75    1       10
#> 44  401112431102976701      2               32          75    1       10
#> 45  401112431102867301      2               26          63    1       10
#> 46  401112431102957801      2               21          75    1       10
#> 47  401112431102957801      2               21          75    1       10
#> 48  401112431102918501      2               14          32    1       10
#> 49  401112431102918501      2               14          32    1       10
#> 50  401112431102978601      2               13          74    2        9
#> 51  401112431102978601      2               13          74    2        9
#> 52  401112431102979101      2                8          75    3       10
#> 53  401112431102959201      2                7          58    1       10
#> 54  401112431102959201      2                7          58    1       10
#> 55  401112431102959201      2                7          58    1       10
#> 56  401112431102959201      2                7          58    1       10
#> 57  401112431102909501      2                4          39    2        7
#> 58  401112431102869601      2                3          52    1       10
#> 59  401112431102869601      2                3          52    1       10
#> 60  401112431103894301      3               56          74    1       10
#> 61  401112431103874701      3               52          63    2        8
#> 62  401112431103874701      3               52          63    2        8
#> 63  401112431103864801      3               51          74    2        9
#> 64  401112431103955001      3               49          47    1       10
#> 65  401112431103875301      3               46          66    3       11
#> 66  401112431103875301      3               46          66    3       11
#> 67  401112431103975501      3               44          75    1       10
#> 68  401112431103975501      3               44          75    1       10
#> 69  401112431103955701      3               42          33    1       10
#> 70  401112431103955701      3               42          33    1       10
#> 71  401112431103865701      3               42          73    3        8
#> 72  401112431103865701      3               42          73    3        8
#> 73  401112431103976301      3               36          75    2       10
#> 74  401112431103976301      3               36          75    2       10
#> 75  401112431103936301      3               36          60    1       10
#> 76  401112431103936301      3               36          60    1       10
#> 77  401112431103937201      3               27          46    1       10
#> 78  401112431103937201      3               27          46    1       10
#> 79  401112431103897301      3               26          77    2       13
#> 80  401112431103867701      3               22          65    1       10
#> 81  401112431103937801      3               21          46    2       10
#> 82  401112431103937801      3               21          46    2       10
#> 83  401112431103928101      3               18          42    3       13
#> 84  401112431103858301      3               16          85    1       20
#> 85  401112431103858301      3               16          85    1       20
#> 86  401112431103978401      3               15          75    3       10
#> 87  401112431103938801      3               11          46    3       10
#> 88  401112431103938801      3               11          46    3       10
#> 89  401112431103899201      3                7          73    3        9
#> 90  401112431103899201      3                7          73    3        9
#> 91  401112431103949701      3                2          48    1       10
#> 92  401112431103949701      3                2          48    1       10
#> 93  401112431104884201      4               57          74    2        9
#> 94  401112431104884201      4               57          74    2        9
#> 95  401112431104984801      4               51          64    1       10
#> 96  401112431104995101      4               48          67    4       13
#> 97  401112431104935301      4               46          74    3        9
#> 98  401112431104975701      4               42          48    3       10
#> 99  401112431104886101      4               38          75    3       10
#> 100 401112431104886101      4               38          75    3       10
#> 101 401112431104986501      4               34          61    2        7
#> 102 401112431104986501      4               34          61    2        7
#> 103 401112431104877101      4               28          75    1       10
#> 104 401112431104988401      4               15          59    3        5
#> 105 401112431104928801      4               11          75    1       10
#> 106 401112431104978901      4               10          75    1       10
#> 107 401112431104978901      4               10          75    1       10
#> 108 401112431104929501      4                4          74    2        9
#> 109 401112431104929501      4                4          74    2        9
#>     athleteId stat            Rush         Fumble              Sack
#> 1     4035628   12            <NA>           <NA>              <NA>
#> 2     4240020   12            <NA>           <NA>              <NA>
#> 3     4240021    0       Cam Akers           <NA>              <NA>
#> 4     4240021    1            <NA>           <NA>              <NA>
#> 5     4240020    1            <NA>           <NA>              <NA>
#> 6     4240020    1            <NA>           <NA>              <NA>
#> 7     4240033   12            <NA>           <NA>              <NA>
#> 8     4240020   12            <NA>           <NA>              <NA>
#> 9     4240028    1            <NA>           <NA>              <NA>
#> 10    4240020    1            <NA>           <NA>              <NA>
#> 11    4240028   75            <NA>           <NA>              <NA>
#> 12    4240020   75            <NA>           <NA>              <NA>
#> 13    4363032   14 Keyshawn Helton           <NA>              <NA>
#> 14    4240034    9            <NA>           <NA>              <NA>
#> 15    4240020    9            <NA>           <NA>              <NA>
#> 16    4240021   10       Cam Akers           <NA>              <NA>
#> 17    4240034   10   D.J. Matthews           <NA>              <NA>
#> 18    4240021    1            <NA>           <NA>              <NA>
#> 19    4240021   38       Cam Akers           <NA>              <NA>
#> 20    4240028    1            <NA>           <NA>              <NA>
#> 21    4240021    1            <NA>           <NA>              <NA>
#> 22    4240021    4       Cam Akers           <NA>              <NA>
#> 23    4240033   13            <NA>           <NA>              <NA>
#> 24    4240020   13            <NA>           <NA>              <NA>
#> 25    4035611    1            <NA>           <NA>              <NA>
#> 26    4240020    1            <NA>           <NA>              <NA>
#> 27    4035611    5            <NA>           <NA>              <NA>
#> 28    4240020    5            <NA>           <NA>              <NA>
#> 29    4035628   14            <NA>           <NA>              <NA>
#> 30    4240020   14            <NA>           <NA>              <NA>
#> 31    4426397    1            <NA>           <NA>      Jaleel McRae
#> 32    4240020    1            <NA>           <NA>              <NA>
#> 33    4240020    1            <NA> James Blackman              <NA>
#> 34    4240020   -2  James Blackman           <NA>              <NA>
#> 35    4240021    3       Cam Akers           <NA>              <NA>
#> 36    4240026    1            <NA>           <NA>    Joshua Kaindoh
#> 37    4240021    5       Cam Akers           <NA>              <NA>
#> 38    4240022   12   Khalan Laborn           <NA>              <NA>
#> 39    4240020    1            <NA> James Blackman              <NA>
#> 40    4240020   15            <NA>           <NA>              <NA>
#> 41    4240022   -1   Khalan Laborn           <NA>              <NA>
#> 42    4240021   31       Cam Akers           <NA>              <NA>
#> 43    4240033    1            <NA>           <NA>              <NA>
#> 44    4240020    1            <NA>           <NA>              <NA>
#> 45    4240020   11            <NA>           <NA>              <NA>
#> 46    4240028   17            <NA>           <NA>              <NA>
#> 47    4240020   17            <NA>           <NA>              <NA>
#> 48    4240034    1            <NA>           <NA>              <NA>
#> 49    4240020    1            <NA>           <NA>              <NA>
#> 50    4363030   -1            <NA>           <NA>              <NA>
#> 51    4240020   -1            <NA>           <NA>              <NA>
#> 52    4240020    5            <NA>           <NA>              <NA>
#> 53    4363032    1            <NA>           <NA>              <NA>
#> 54    4240020    1            <NA>           <NA>              <NA>
#> 55    4363032   58            <NA>           <NA>              <NA>
#> 56    4240020   58            <NA>           <NA>              <NA>
#> 57    4363036    1            <NA>           <NA>              <NA>
#> 58    4363030   51            <NA>           <NA>              <NA>
#> 59    4240020   51            <NA>           <NA>              <NA>
#> 60    4240022   -3   Khalan Laborn           <NA>              <NA>
#> 61    4240034   -3            <NA>           <NA>              <NA>
#> 62    4240020   -3            <NA>           <NA>              <NA>
#> 63    4240021    1       Cam Akers           <NA>              <NA>
#> 64    4240021   14       Cam Akers           <NA>              <NA>
#> 65    4240033    1            <NA>           <NA>              <NA>
#> 66    4240020    1            <NA>           <NA>              <NA>
#> 67    4240028    1            <NA>           <NA>              <NA>
#> 68    4240020    1            <NA>           <NA>              <NA>
#> 69    4240021    1            <NA>      Cam Akers              <NA>
#> 70    4240021    4       Cam Akers           <NA>              <NA>
#> 71    4240023    8            <NA>           <NA>              <NA>
#> 72    4240020    8            <NA>           <NA>              <NA>
#> 73    4363032    1            <NA>           <NA>              <NA>
#> 74    4240020    1            <NA>           <NA>              <NA>
#> 75    4240034   14            <NA>           <NA>              <NA>
#> 76    4240020   14            <NA>           <NA>              <NA>
#> 77    4035628    1            <NA>           <NA>              <NA>
#> 78    4240020    1            <NA>           <NA>              <NA>
#> 79    4240020    4  James Blackman           <NA>              <NA>
#> 80    4240021    2       Cam Akers           <NA>              <NA>
#> 81    4240034    1            <NA>           <NA>              <NA>
#> 82    4240020    1            <NA>           <NA>              <NA>
#> 83    4035605    1            <NA>           <NA> Carlos Becker III
#> 84    4240021   11            <NA>           <NA>              <NA>
#> 85    4240020   11            <NA>           <NA>              <NA>
#> 86    4240020    9            <NA>           <NA>              <NA>
#> 87    4240028   -1            <NA>           <NA>              <NA>
#> 88    4240020   -1            <NA>           <NA>              <NA>
#> 89    4240033    1            <NA>           <NA>              <NA>
#> 90    4240020    1            <NA>           <NA>              <NA>
#> 91    4363047    1            <NA>           <NA>              <NA>
#> 92    4240029    1            <NA>           <NA>     Marvin Wilson
#> 93    4240034   -1            <NA>           <NA>              <NA>
#> 94    4240020   -1            <NA>           <NA>              <NA>
#> 95    4240021    3       Cam Akers           <NA>              <NA>
#> 96    4240020    9  James Blackman           <NA>              <NA>
#> 97    4240020    7            <NA>           <NA>              <NA>
#> 98    4035605    1            <NA>           <NA> Carlos Becker III
#> 99    4240028    8            <NA>           <NA>              <NA>
#> 100   4240020    8            <NA>           <NA>              <NA>
#> 101   4035628    2            <NA>           <NA>              <NA>
#> 102   4240020    2            <NA>           <NA>              <NA>
#> 103   4240021    1       Cam Akers           <NA>              <NA>
#> 104   4240021    1       Cam Akers           <NA>              <NA>
#> 105   4240021    1       Cam Akers           <NA>              <NA>
#> 106   4035628   11            <NA>           <NA>              <NA>
#> 107   4240020   11            <NA>           <NA>              <NA>
#> 108   4240028    1            <NA>           <NA>              <NA>
#> 109   4240020    1            <NA>           <NA>              <NA>
#>           Touchdown   Incompletion       Reception     Completion
#> 1              <NA>           <NA>     Keith Gavin           <NA>
#> 2              <NA>           <NA>            <NA> James Blackman
#> 3              <NA>           <NA>            <NA>           <NA>
#> 4              <NA>           <NA>       Cam Akers           <NA>
#> 5              <NA>           <NA>            <NA> James Blackman
#> 6              <NA> James Blackman            <NA>           <NA>
#> 7              <NA>           <NA>  Ontaria Wilson           <NA>
#> 8              <NA>           <NA>            <NA> James Blackman
#> 9   Tamorrion Terry           <NA>            <NA>           <NA>
#> 10   James Blackman           <NA>            <NA>           <NA>
#> 11             <NA>           <NA> Tamorrion Terry           <NA>
#> 12             <NA>           <NA>            <NA> James Blackman
#> 13             <NA>           <NA>            <NA>           <NA>
#> 14             <NA>           <NA>   D.J. Matthews           <NA>
#> 15             <NA>           <NA>            <NA> James Blackman
#> 16             <NA>           <NA>            <NA>           <NA>
#> 17             <NA>           <NA>            <NA>           <NA>
#> 18        Cam Akers           <NA>            <NA>           <NA>
#> 19             <NA>           <NA>            <NA>           <NA>
#> 20             <NA>           <NA>            <NA>           <NA>
#> 21             <NA>      Cam Akers            <NA>           <NA>
#> 22             <NA>           <NA>            <NA>           <NA>
#> 23             <NA>           <NA>  Ontaria Wilson           <NA>
#> 24             <NA>           <NA>            <NA> James Blackman
#> 25      Gabe Nabers           <NA>            <NA>           <NA>
#> 26   James Blackman           <NA>            <NA>           <NA>
#> 27             <NA>           <NA>     Gabe Nabers           <NA>
#> 28             <NA>           <NA>            <NA> James Blackman
#> 29             <NA>           <NA>     Keith Gavin           <NA>
#> 30             <NA>           <NA>            <NA> James Blackman
#> 31             <NA>           <NA>            <NA>           <NA>
#> 32             <NA> James Blackman            <NA>           <NA>
#> 33             <NA>           <NA>            <NA>           <NA>
#> 34             <NA>           <NA>            <NA>           <NA>
#> 35             <NA>           <NA>            <NA>           <NA>
#> 36             <NA>           <NA>            <NA>           <NA>
#> 37             <NA>           <NA>            <NA>           <NA>
#> 38             <NA>           <NA>            <NA>           <NA>
#> 39             <NA>           <NA>            <NA>           <NA>
#> 40             <NA>           <NA>            <NA>           <NA>
#> 41             <NA>           <NA>            <NA>           <NA>
#> 42             <NA>           <NA>            <NA>           <NA>
#> 43             <NA>           <NA>  Ontaria Wilson           <NA>
#> 44             <NA>           <NA>            <NA> James Blackman
#> 45             <NA>           <NA>            <NA> James Blackman
#> 46             <NA>           <NA> Tamorrion Terry           <NA>
#> 47             <NA>           <NA>            <NA> James Blackman
#> 48             <NA>           <NA>            <NA>           <NA>
#> 49             <NA> James Blackman            <NA>           <NA>
#> 50             <NA>           <NA> Warren Thompson           <NA>
#> 51             <NA>           <NA>            <NA> James Blackman
#> 52             <NA>           <NA>            <NA>           <NA>
#> 53  Keyshawn Helton           <NA>            <NA>           <NA>
#> 54   James Blackman           <NA>            <NA>           <NA>
#> 55             <NA>           <NA> Keyshawn Helton           <NA>
#> 56             <NA>           <NA>            <NA> James Blackman
#> 57             <NA>           <NA>            <NA>           <NA>
#> 58             <NA>           <NA> Warren Thompson           <NA>
#> 59             <NA>           <NA>            <NA> James Blackman
#> 60             <NA>           <NA>            <NA>           <NA>
#> 61             <NA>           <NA>   D.J. Matthews           <NA>
#> 62             <NA>           <NA>            <NA> James Blackman
#> 63             <NA>           <NA>            <NA>           <NA>
#> 64             <NA>           <NA>            <NA>           <NA>
#> 65             <NA>           <NA>            <NA>           <NA>
#> 66             <NA> James Blackman            <NA>           <NA>
#> 67             <NA>           <NA>            <NA>           <NA>
#> 68             <NA> James Blackman            <NA>           <NA>
#> 69             <NA>           <NA>            <NA>           <NA>
#> 70             <NA>           <NA>            <NA>           <NA>
#> 71             <NA>           <NA>    Tre' McKitty           <NA>
#> 72             <NA>           <NA>            <NA> James Blackman
#> 73             <NA>           <NA>            <NA>           <NA>
#> 74             <NA> James Blackman            <NA>           <NA>
#> 75             <NA>           <NA>   D.J. Matthews           <NA>
#> 76             <NA>           <NA>            <NA> James Blackman
#> 77             <NA>           <NA>            <NA>           <NA>
#> 78             <NA> James Blackman            <NA>           <NA>
#> 79             <NA>           <NA>            <NA>           <NA>
#> 80             <NA>           <NA>            <NA>           <NA>
#> 81             <NA>           <NA>            <NA>           <NA>
#> 82             <NA> James Blackman            <NA>           <NA>
#> 83             <NA>           <NA>            <NA>           <NA>
#> 84             <NA>           <NA>       Cam Akers           <NA>
#> 85             <NA>           <NA>            <NA> James Blackman
#> 86             <NA>           <NA>            <NA>           <NA>
#> 87             <NA>           <NA> Tamorrion Terry           <NA>
#> 88             <NA>           <NA>            <NA> James Blackman
#> 89             <NA>           <NA>            <NA>           <NA>
#> 90             <NA> James Blackman            <NA>           <NA>
#> 91             <NA>           <NA>            <NA>           <NA>
#> 92             <NA>           <NA>            <NA>           <NA>
#> 93             <NA>           <NA>   D.J. Matthews           <NA>
#> 94             <NA>           <NA>            <NA> James Blackman
#> 95             <NA>           <NA>            <NA>           <NA>
#> 96             <NA>           <NA>            <NA>           <NA>
#> 97             <NA>           <NA>            <NA>           <NA>
#> 98             <NA>           <NA>            <NA>           <NA>
#> 99             <NA>           <NA> Tamorrion Terry           <NA>
#> 100            <NA>           <NA>            <NA> James Blackman
#> 101            <NA>           <NA>     Keith Gavin           <NA>
#> 102            <NA>           <NA>            <NA> James Blackman
#> 103            <NA>           <NA>            <NA>           <NA>
#> 104            <NA>           <NA>            <NA>           <NA>
#> 105            <NA>           <NA>            <NA>           <NA>
#> 106            <NA>           <NA>     Keith Gavin           <NA>
#> 107            <NA>           <NA>            <NA> James Blackman
#> 108            <NA>           <NA>            <NA>           <NA>
#> 109            <NA> James Blackman            <NA>           <NA>
#>              Target     Sack Taken Fumble Recovered Fumble Forced  Interception
#> 1              <NA>           <NA>             <NA>          <NA>          <NA>
#> 2              <NA>           <NA>             <NA>          <NA>          <NA>
#> 3              <NA>           <NA>             <NA>          <NA>          <NA>
#> 4              <NA>           <NA>             <NA>          <NA>          <NA>
#> 5              <NA>           <NA>             <NA>          <NA>          <NA>
#> 6              <NA>           <NA>             <NA>          <NA>          <NA>
#> 7              <NA>           <NA>             <NA>          <NA>          <NA>
#> 8              <NA>           <NA>             <NA>          <NA>          <NA>
#> 9              <NA>           <NA>             <NA>          <NA>          <NA>
#> 10             <NA>           <NA>             <NA>          <NA>          <NA>
#> 11             <NA>           <NA>             <NA>          <NA>          <NA>
#> 12             <NA>           <NA>             <NA>          <NA>          <NA>
#> 13             <NA>           <NA>             <NA>          <NA>          <NA>
#> 14             <NA>           <NA>             <NA>          <NA>          <NA>
#> 15             <NA>           <NA>             <NA>          <NA>          <NA>
#> 16             <NA>           <NA>             <NA>          <NA>          <NA>
#> 17             <NA>           <NA>             <NA>          <NA>          <NA>
#> 18             <NA>           <NA>             <NA>          <NA>          <NA>
#> 19             <NA>           <NA>             <NA>          <NA>          <NA>
#> 20  Tamorrion Terry           <NA>             <NA>          <NA>          <NA>
#> 21             <NA>           <NA>             <NA>          <NA>          <NA>
#> 22             <NA>           <NA>             <NA>          <NA>          <NA>
#> 23             <NA>           <NA>             <NA>          <NA>          <NA>
#> 24             <NA>           <NA>             <NA>          <NA>          <NA>
#> 25             <NA>           <NA>             <NA>          <NA>          <NA>
#> 26             <NA>           <NA>             <NA>          <NA>          <NA>
#> 27             <NA>           <NA>             <NA>          <NA>          <NA>
#> 28             <NA>           <NA>             <NA>          <NA>          <NA>
#> 29             <NA>           <NA>             <NA>          <NA>          <NA>
#> 30             <NA>           <NA>             <NA>          <NA>          <NA>
#> 31             <NA>           <NA>             <NA>          <NA>          <NA>
#> 32             <NA>           <NA>             <NA>          <NA>          <NA>
#> 33             <NA>           <NA>             <NA>          <NA>          <NA>
#> 34             <NA>           <NA>             <NA>          <NA>          <NA>
#> 35             <NA>           <NA>             <NA>          <NA>          <NA>
#> 36             <NA>           <NA>             <NA>          <NA>          <NA>
#> 37             <NA>           <NA>             <NA>          <NA>          <NA>
#> 38             <NA>           <NA>             <NA>          <NA>          <NA>
#> 39             <NA>           <NA>             <NA>          <NA>          <NA>
#> 40             <NA> James Blackman             <NA>          <NA>          <NA>
#> 41             <NA>           <NA>             <NA>          <NA>          <NA>
#> 42             <NA>           <NA>             <NA>          <NA>          <NA>
#> 43             <NA>           <NA>             <NA>          <NA>          <NA>
#> 44             <NA>           <NA>             <NA>          <NA>          <NA>
#> 45             <NA>           <NA>             <NA>          <NA>          <NA>
#> 46             <NA>           <NA>             <NA>          <NA>          <NA>
#> 47             <NA>           <NA>             <NA>          <NA>          <NA>
#> 48    D.J. Matthews           <NA>             <NA>          <NA>          <NA>
#> 49             <NA>           <NA>             <NA>          <NA>          <NA>
#> 50             <NA>           <NA>             <NA>          <NA>          <NA>
#> 51             <NA>           <NA>             <NA>          <NA>          <NA>
#> 52             <NA> James Blackman             <NA>          <NA>          <NA>
#> 53             <NA>           <NA>             <NA>          <NA>          <NA>
#> 54             <NA>           <NA>             <NA>          <NA>          <NA>
#> 55             <NA>           <NA>             <NA>          <NA>          <NA>
#> 56             <NA>           <NA>             <NA>          <NA>          <NA>
#> 57             <NA>           <NA>             <NA>          <NA> Isaiah Bolden
#> 58             <NA>           <NA>             <NA>          <NA>          <NA>
#> 59             <NA>           <NA>             <NA>          <NA>          <NA>
#> 60             <NA>           <NA>             <NA>          <NA>          <NA>
#> 61             <NA>           <NA>             <NA>          <NA>          <NA>
#> 62             <NA>           <NA>             <NA>          <NA>          <NA>
#> 63             <NA>           <NA>             <NA>          <NA>          <NA>
#> 64             <NA>           <NA>             <NA>          <NA>          <NA>
#> 65   Ontaria Wilson           <NA>             <NA>          <NA>          <NA>
#> 66             <NA>           <NA>             <NA>          <NA>          <NA>
#> 67  Tamorrion Terry           <NA>             <NA>          <NA>          <NA>
#> 68             <NA>           <NA>             <NA>          <NA>          <NA>
#> 69             <NA>           <NA>             <NA>          <NA>          <NA>
#> 70             <NA>           <NA>             <NA>          <NA>          <NA>
#> 71             <NA>           <NA>             <NA>          <NA>          <NA>
#> 72             <NA>           <NA>             <NA>          <NA>          <NA>
#> 73  Keyshawn Helton           <NA>             <NA>          <NA>          <NA>
#> 74             <NA>           <NA>             <NA>          <NA>          <NA>
#> 75             <NA>           <NA>             <NA>          <NA>          <NA>
#> 76             <NA>           <NA>             <NA>          <NA>          <NA>
#> 77      Keith Gavin           <NA>             <NA>          <NA>          <NA>
#> 78             <NA>           <NA>             <NA>          <NA>          <NA>
#> 79             <NA>           <NA>             <NA>          <NA>          <NA>
#> 80             <NA>           <NA>             <NA>          <NA>          <NA>
#> 81    D.J. Matthews           <NA>             <NA>          <NA>          <NA>
#> 82             <NA>           <NA>             <NA>          <NA>          <NA>
#> 83             <NA>           <NA>             <NA>          <NA>          <NA>
#> 84             <NA>           <NA>             <NA>          <NA>          <NA>
#> 85             <NA>           <NA>             <NA>          <NA>          <NA>
#> 86             <NA> James Blackman             <NA>          <NA>          <NA>
#> 87             <NA>           <NA>             <NA>          <NA>          <NA>
#> 88             <NA>           <NA>             <NA>          <NA>          <NA>
#> 89   Ontaria Wilson           <NA>             <NA>          <NA>          <NA>
#> 90             <NA>           <NA>             <NA>          <NA>          <NA>
#> 91             <NA>           <NA>    Robert Cooper Robert Cooper          <NA>
#> 92             <NA>           <NA>             <NA>          <NA>          <NA>
#> 93             <NA>           <NA>             <NA>          <NA>          <NA>
#> 94             <NA>           <NA>             <NA>          <NA>          <NA>
#> 95             <NA>           <NA>             <NA>          <NA>          <NA>
#> 96             <NA>           <NA>             <NA>          <NA>          <NA>
#> 97             <NA> James Blackman             <NA>          <NA>          <NA>
#> 98             <NA>           <NA>             <NA>          <NA>          <NA>
#> 99             <NA>           <NA>             <NA>          <NA>          <NA>
#> 100            <NA>           <NA>             <NA>          <NA>          <NA>
#> 101            <NA>           <NA>             <NA>          <NA>          <NA>
#> 102            <NA>           <NA>             <NA>          <NA>          <NA>
#> 103            <NA>           <NA>             <NA>          <NA>          <NA>
#> 104            <NA>           <NA>             <NA>          <NA>          <NA>
#> 105            <NA>           <NA>             <NA>          <NA>          <NA>
#> 106            <NA>           <NA>             <NA>          <NA>          <NA>
#> 107            <NA>           <NA>             <NA>          <NA>          <NA>
#> 108 Tamorrion Terry           <NA>             <NA>          <NA>          <NA>
#> 109            <NA>           <NA>             <NA>          <NA>          <NA>