Lists all teams in conference or all D-I teams if conference is left NULL Current support only for D-I

Lists all teams in conference or all D-I teams if conference is left NULL Current support only for D-I

cfb_team(conference = NULL, only_fbs = TRUE, year = NULL)

Arguments

conference
(_String_ optional): Conference abbreviation - Select a valid FBS conference Conference abbreviations P5: ACC, B12, B1G, SEC, PAC, Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

only_fbs
(_Logical_, default TRUE): Filter for only returning FBS teams for a given year. If year is left blank while only_fbs is TRUE, then will return values for most current year

year
(_Integer_ optional): Year, 4 digit format (YYYY). Filter for getting a list of major division team for a given year

Examples


cfb_team(conference = "SEC")

#>      id            school       mascot abbreviation       alt_name1 alt_name2
#> 1   333           Alabama Crimson Tide          ALA            <NA>       ALA
#> 2     8          Arkansas   Razorbacks          ARK            <NA>       ARK
#> 3     2            Auburn       Tigers          AUB            <NA>       AUB
#> 4    57           Florida       Gators          FLA            <NA>       FLA
#> 5    61           Georgia     Bulldogs          UGA            <NA>       UGA
#> 6    96          Kentucky     Wildcats           UK            <NA>        UK
#> 7    99               LSU       Tigers          LSU Louisiana State       LSU
#> 8   344 Mississippi State     Bulldogs         MSST            <NA>      MSST
#> 9   142          Missouri       Tigers          MIZ            <NA>       MIZ
#> 10  145          Ole Miss       Rebels         MISS            <NA>      MISS
#> 11 2579    South Carolina    Gamecocks           SC            <NA>        SC
#> 12 2633         Tennessee   Volunteers         TENN            <NA>      TENN
#> 13  245         Texas A&M       Aggies         TA&M            <NA>      TA&M
#> 14  238        Vanderbilt   Commodores          VAN            <NA>       VAN
#>         alt_name3 conference division   color alt_color
#> 1         Alabama        SEC     West #690014   #f1f2f3
#> 2        Arkansas        SEC     West #9c1831   #000000
#> 3          Auburn        SEC     West #03244d   #f1f2f3
#> 4         Florida        SEC     East #0021A5   #0021a5
#> 5         Georgia        SEC     East #CC0000   #000000
#> 6        Kentucky        SEC     East #005DAA   #ffffff
#> 7             LSU        SEC     West #22005c   #fdd023
#> 8         Miss St        SEC     West #762123   #c8c8c8
#> 9        Missouri        SEC     East #000000   #000000
#> 10       Ole Miss        SEC     West #001148   #00205b
#> 11 South Carolina        SEC     East #670010   #000000
#> 12      Tennessee        SEC     East #EE9627   #ffffff
#> 13      Texas A&M        SEC     West #5C0025   #ffffff
#> 14     Vanderbilt        SEC     East #000000   #231f20
#>                                                                                                          logos
#> 1    http://a.espncdn.com/i/teamlogos/ncaa/500/333.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/333.png
#> 2        http://a.espncdn.com/i/teamlogos/ncaa/500/8.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/8.png
#> 3        http://a.espncdn.com/i/teamlogos/ncaa/500/2.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2.png
#> 4      http://a.espncdn.com/i/teamlogos/ncaa/500/57.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/57.png
#> 5      http://a.espncdn.com/i/teamlogos/ncaa/500/61.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/61.png
#> 6      http://a.espncdn.com/i/teamlogos/ncaa/500/96.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/96.png
#> 7      http://a.espncdn.com/i/teamlogos/ncaa/500/99.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/99.png
#> 8    http://a.espncdn.com/i/teamlogos/ncaa/500/344.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/344.png
#> 9    http://a.espncdn.com/i/teamlogos/ncaa/500/142.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/142.png
#> 10   http://a.espncdn.com/i/teamlogos/ncaa/500/145.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/145.png
#> 11 http://a.espncdn.com/i/teamlogos/ncaa/500/2579.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2579.png
#> 12 http://a.espncdn.com/i/teamlogos/ncaa/500/2633.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2633.png
#> 13   http://a.espncdn.com/i/teamlogos/ncaa/500/245.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/245.png
#> 14   http://a.espncdn.com/i/teamlogos/ncaa/500/238.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/238.png


cfb_team(conference = "Ind")

#>     id           school         mascot abbreviation     alt_name1 alt_name2
#> 1  349             Army  Black Knights         ARMY          <NA>      ARMY
#> 2  252              BYU        Cougars          BYU Brigham Young       BYU
#> 3 2335          Liberty         Flames          LIB          <NA>       LIB
#> 4  166 New Mexico State         Aggies         NMSU          <NA>      NMSU
#> 5   87       Notre Dame Fighting Irish           ND          <NA>        ND
#> 6  113            UMass      Minutemen        UMASS          <NA>     UMASS
#>       alt_name3       conference division   color alt_color
#> 1          Army FBS Independents       NA #ce9c00   #231f20
#> 2           BYU FBS Independents       NA #001E4C   #ffffff
#> 3       Liberty FBS Independents       NA #071740   #a61f21
#> 4 New Mexico St FBS Independents       NA #891216   #000000
#> 5    Notre Dame FBS Independents       NA #00122b   #ae9142
#> 6         UMass FBS Independents       NA #880007      <NA>
#>                                                                                                         logos
#> 1   http://a.espncdn.com/i/teamlogos/ncaa/500/349.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/349.png
#> 2   http://a.espncdn.com/i/teamlogos/ncaa/500/252.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/252.png
#> 3 http://a.espncdn.com/i/teamlogos/ncaa/500/2335.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2335.png
#> 4   http://a.espncdn.com/i/teamlogos/ncaa/500/166.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/166.png
#> 5     http://a.espncdn.com/i/teamlogos/ncaa/500/87.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/87.png
#> 6   http://a.espncdn.com/i/teamlogos/ncaa/500/113.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/113.png