Team Info Lookup Lists all teams in conference or all D-I teams if conference is left NULL Current support only for D-I

Team Info Lookup Lists all teams in conference or all D-I teams if conference is left NULL Current support only for D-I

cfb_team_info(conference = NULL, only_fbs = TRUE, year = NULL)

Arguments

conference
(String optional): Conference abbreviation - Select a valid FBS conference Conference abbreviations P5: ACC, B12, B1G, SEC, PAC, Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

only_fbs
(Logical default TRUE): Filter for only returning FBS teams for a given year. If year is left blank while only_fbs is TRUE, then will return values for most current year

year
(Integer optional): Year, 4 digit format (YYYY). Filter for getting a list of major division team for a given year

Examples


cfb_team_info(conference = "SEC")

#> Error in cfb_team_info(conference = "SEC"): could not find function "cfb_team_info"


cfb_team_info(conference = "Ind")

#> Error in cfb_team_info(conference = "Ind"): could not find function "cfb_team_info"


cfb_team_info(year = 2019)

#> Error in cfb_team_info(year = 2019): could not find function "cfb_team_info"